### 2018-09-07 #932 @arfon
Hi @arvestad - please review the [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html) and prepare a paper. The review cannot proceed until you have done this.
