### 2018-09-07 303ab5d
Preparing for JOSS
