+++ doc/joss/paper.bib
@@ -1,0 +1,127 @@
+@Misc{arvestad18,
+  author = 	 {Lars Arvestad},
+  title = 	 {alv: a console-based alignment viewer},
+  howpublished = {https://github.com/arvestad/alv},
+  year = 	 2018}
+
+@article{waterhouse2009jalview,
+  title={Jalview Version 2—a multiple sequence alignment editor and analysis workbench},
+  author={Waterhouse, Andrew M and Procter, James B and Martin, David MA and Clamp, Mich{\`e}le and Barton, Geoffrey J},
+  journal={Bioinformatics},
+  volume={25},
+  number={9},
+  pages={1189--1191},
+  year={2009},
+  doi = {10.1093/bioinformatics/btp033},
+  publisher={Oxford University Press}
+}
+@article{gouy2009seaview,
+  title={SeaView version 4: a multiplatform graphical user interface for sequence alignment and phylogenetic tree building},
+  author={Gouy, Manolo and Guindon, St{\'e}phane and Gascuel, Olivier},
+  journal={Molecular biology and evolution},
+  volume={27},
+  number={2},
+  pages={221--224},
+  year={2009},
+  doi = {10.1093/molbev/msp259},
+  publisher={Oxford University Press}
+}
+
+@online{belvu,
+  author = 	 {Erik Sonnhammer},
+  title = 	 {Belvu},
+  url = {http://sonnhammer.sbc.su.se/Belvu.html}}
+
+@article{ranwez2011macse,
+  title={MACSE: Multiple Alignment of Coding SEquences accounting for frameshifts and stop codons},
+  author={Ranwez, Vincent and Harispe, S{\'e}bastien and Delsuc, Fr{\'e}d{\'e}ric and Douzery, Emmanuel JP},
+  journal={PloS ONE},
+  volume={6},
+  number={9},
+  pages={e22594},
+  year={2011},
+  doi = {10.1371/journal.pone.0022594},
+  publisher={Public Library of Science}
+}
+
+@incollection{sievers2014clustal,
+  title={Clustal Omega, accurate alignment of very large numbers of sequences},
+  author={Sievers, Fabian and Higgins, Desmond G},
+  booktitle={Multiple sequence alignment methods},
+  pages={105--116},
+  year={2014},
+  doi = {10.1002/0471250953.bi0313s48},
+  publisher={Springer}
+}
+@article{katoh2013mafft,
+  title={MAFFT multiple sequence alignment software version 7: improvements in performance and usability},
+  author={Katoh, Kazutaka and Standley, Daron M},
+  journal={Molecular biology and evolution},
+  volume={30},
+  number={4},
+  pages={772--780},
+  year={2013},
+  doi = {10.1093/molbev/mst010},
+  publisher={Society for Molecular Biology and Evolution}
+}
+@article{edgar2004muscle,
+  title={MUSCLE: multiple sequence alignment with high accuracy and high throughput},
+  author={Edgar, Robert C},
+  journal={Nucleic acids research},
+  volume={32},
+  number={5},
+  pages={1792--1797},
+  year={2004},
+  doi = {10.1186/1471-2105-5-113},
+  publisher={Oxford University Press}
+}
+@online{hmmer,
+  author = 	 {Sean Eddy},
+  title = 	 {hmmalign: align sequences to a profile HMM},
+  url = {hmmer.org},
+  year = 	 2015}
+
+@article{veidenberg2015wasabi,
+  title={Wasabi: an integrated platform for evolutionary sequence analysis and data visualization},
+  author={Veidenberg, Andres and Medlar, Alan and L{\"o}ytynoja, Ari},
+  journal={Molecular biology and evolution},
+  volume={33},
+  number={4},
[…]
