### 2018-09-07 #932 @arvestad
I am on it. I had forgotten to commit/push the prepared paper to the repo. :-(

	Thanks,
	L

> On 7 Sep 2018, at 13:54, Arfon Smith <notifications@github.com> wrote:
>
> Hi @arvestad <https://github.com/arvestad> - please review the submission guidelines <https://joss.readthedocs.io/en/latest/submitting.html> and prepare a paper. The review cannot proceed until you have done this.
>
