### 2018-09-07 d7b6724
Added two forgotten refs and fixed capitalisation
