+++ doc/joss/paper.bib
@@ -32,6 +32,18 @@
   title = 	 {Belvu},
   url = {http://sonnhammer.sbc.su.se/Belvu.html}}
 
+@online{ncbimsaviewer,
+  publisher={NCBI},
+  title={NCBI Multiple Sequence Alignment Viewer},
+  url={https://www.ncbi.nlm.nih.gov/projects/msaviewer/}
+}
+
+@online{mview,
+                  title={MView},
+url={https://www.ebi.ac.uk/Tools/msa/mview},
+ publisher={EBI}
+}
+
 @article{ranwez2011macse,
   title={MACSE: Multiple Alignment of Coding SEquences accounting for frameshifts and stop codons},
   author={Ranwez, Vincent and Harispe, S{\'e}bastien and Delsuc, Fr{\'e}d{\'e}ric and Douzery, Emmanuel JP},
@@ -77,7 +89,7 @@
 }
 @online{hmmer,
   author = 	 {Sean Eddy},
-  title = 	 {hmmalign: align sequences to a profile HMM},
+  title = 	 {{hmmalign}: align sequences to a profile {HMM}},
   url = {hmmer.org},
   year = 	 2015}
 
