### 2018-09-07 d980367
Removed a redundant word
