+++ doc/joss/paper.bib
@@ -32,15 +32,15 @@
   title = 	 {Belvu},
   url = {http://sonnhammer.sbc.su.se/Belvu.html}}
 
-@online{ncbimsaviewer,
+@misc{ncbimsaviewer,
   publisher={NCBI},
   title={NCBI Multiple Sequence Alignment Viewer},
   url={https://www.ncbi.nlm.nih.gov/projects/msaviewer/}
 }
 
-@online{mview,
-                  title={MView},
-url={https://www.ebi.ac.uk/Tools/msa/mview},
+@misc{mview,
+ title={MView},
+ url={https://www.ebi.ac.uk/Tools/msa/mview},
  publisher={EBI}
 }
 
