### 2018-09-07 #932 @arvestad
My suggestions for reviewers:

* Karin Lagesen, @karinlag. Involved in Software Carpentry and advocate for quality in Bioinformatics software.

I picked three names with an interest in evolution from the list of potential referees:

* @giraola
* @shaunpwilkinson
* @juanvillada

Bioinformatics is a broad subject these days, but people working in evolution are more likely to be interested in using this software.
