### 2018-09-16 #955 @whedon
[REVIEW]: alv: a console-based viewer for molecular sequence alignments
**Submitting author:** @arvestad (<a href="http://orcid.org/0000-0001-5341-1733">Lars Arvestad</a>)
**Repository:** <a href="https://github.com/arvestad/alv" target ="_blank">https://github.com/arvestad/alv</a>
**Version:** v1.2.0
