### 2018-09-18 #955 @juanvillada
Hi @arvestad, this is a very nice and useful tool!

Some minor suggestions:

1. Installation proceeded smoothly. I just specified a `pip` update to make sure it is updated before installing `alv`:

```sh
pip install --upgrade pip
pip install alv
```
I would suggest you to consider moving the [installation](https://github.com/arvestad/alv#install) section to the top, before the [examples](https://github.com/arvestad/alv#examples).

2. The examples seem enough to show the functions of `alv`. However, in the [examples](https://github.com/arvestad/alv#examples) section appears:

```sh
alf msa.fa
```
Where I suppose it should be:

```sh
alv msa.fa
```

Same for:

```sh
alf -so a,b,c msa.fa
```

3. You may consider moving one of the [screenshots](https://github.com/arvestad/alv#screenshot) to the [introduction section](https://github.com/arvestad/alv#alv-a-command-line-alignment-viewer) in order to call the attention of users.

I already tested `alv` on macOS and will test it on centOS later this week. I will come back with some other comments after that.

Best wishes!
