### 2018-09-21 a038c3f
Tried to avoid repeated length calculations
