### 2018-09-21 c2ff508
Fixed errors and improved presentation, as suggested by reviewer \@juanvilada
