### 2018-09-25 #955 @arvestad
Thanks for the nice comments @juanvillada ! I have adapted your suggestions.
