+++ THANKS.md
@@ -1,0 +1,4 @@
+# Project support
+
+* @katieemalianova help chase down installation issues.
+* @SimonGreenhill added support for the Nexus format.
