### 2018-10-02 #955 @karinlag
I have now finished my review.

First, @arvestad what a nice little tool!

Install instructions are nice and easy to follow, and it is easy to use and understand. The paper also clearly demonstrates the need, and it seems like this software fills that need.

I'd like to see a CONTRIBUTING.md file, detailing how to contribute, otherwise this looks good to go for me.
