### 2018-10-03 #955 @juanvillada
Hi @arvestad

Some minor typos to correct in README.md:

1. In the [Introduction](https://github.com/arvestad/alv#alv-a-command-line-alignment-viewer) section, it reads:
>View you**r** DNA or protein.

2. In the next line, it reads:
>but this ha~~ve~~**s** not been tested

`alv` really works nice, I am now using it constantly for visual inspection of codon variants. Thanks @arvestad for taking the time to contribute open source software to the world!

@yochannah, I think I am ready to suggest `alv` for publication once @arvestad follows @karinlag comments.
