### 2018-10-03 c618b92
Some hints about contributing, as suggested by Karin L
