+++ CONTRIBUTING.md
@@ -30,7 +30,7 @@ to "the next level", then consider forking this project, rename, and start somet
 
 Please consider creating an issue in the `alv` repo if you have a bug report or feature
 request. Emails are welcome too, but are harder to keep track of in relation to `alv`
-development. GitHub has a guide called ![Mastering
+development. GitHub has a guide called [Mastering
 Issues](https://guides.github.com/features/issues/) to get you started. It is easier than it may
 look!
 
@@ -49,7 +49,7 @@ additions --- anything that might improve `alv` usage.
     - fork this repo on github.com, 
 	- work in your own copy, 
 	- create a pull request to the original alv repo. 
-  See GitHub's guide ![Forking Projects](https://guides.github.com/activities/forking/). 
+  See GitHub's guide [Forking Projects](https://guides.github.com/activities/forking/). 
 * Please try to add a unit test for a contributed feature or fixed bug.
 * For large changes, please discuss with @arvestad first. Or just fork, rename, and continue! This
   is Free Software.
