### 2018-10-03 a7dfd58
Added notice of THANKS.md
