### 2018-10-03 c55c61c
Fixed typos
