### 2018-10-03 c76342d
The tick-mark on non-UTF-8 systems is now, hopefully, a circumflex
