+++ tests/test_io.py
@@ -107,10 +107,16 @@ class TestFormats(unittest.TestCase):
 
 
 class TestIndexBar(unittest.TestCase):
+    def setUp(self):
+        self.tickmark = '^'
+        import sys
+        if sys.stdout.encoding == 'UTF-8':
+            self.tickmark = '↑'
+
     def test_make_one_tick(self):
-        s = at.make_one_tick(5, 10)
+        s = at.make_one_tick(5, 10, '^')
         self.assertEqual(len(s), 10)
-        self.assertEqual(s, '        5↑')
+        self.assertEqual(s, '        5^')
 
     def test_calc_tick_indices(self):
         indices = list(at.calc_tick_indices(33, 100, 10, 7))
@@ -121,16 +127,28 @@ class TestIndexBar(unittest.TestCase):
 
     def test_make_tick_string(self):
         s = at.make_tick_string(6, 0, 40, 40, 10)
-        self.assertEqual(s, '     0↑')
+        if self.tickmark == '↑':
+            self.assertEqual(s, '     0↑')
+        else:
+            self.assertEqual(s, '     0^')
 
         s = at.make_tick_string(6, 0, 40, 20, 10)
-        self.assertEqual(s, '     0↑                 20↑')
+        if self.tickmark == '↑':
+            self.assertEqual(s, '     0↑                 20↑')
+        else:
+            self.assertEqual(s, '     0^                 20^')
 
         s = at.make_tick_string(5, 0, 50, 20, 7)
-        self.assertEqual(s, '    0↑                 20↑                 40↑')
+        if self.tickmark == '↑':
+            self.assertEqual(s, '    0↑                 20↑                 40↑')
+        else:
+            self.assertEqual(s, '    0^                 20^                 40^')
                             
         s = at.make_tick_string(7, 35, 100, 20, 10)
-        self.assertEqual(s, '     35↑                      60↑                 80↑')
+        if self.tickmark == '↑':
+            self.assertEqual(s, '     35↑                      60↑                 80↑')
+        else:
+            self.assertEqual(s, '     35^                      60^                 80^')
 
 
 if __name__ == '__main__':
