### 2018-10-03 #955 @arvestad
Thanks @juanvillada and @karinlag for your nice comments!

I have now created CONTRIBUTING.md file, that nicely complements a file THANKS.md that I started just two days ago, and addressed Juan's minor issues. So I think that I have taken care of the points that have been raised. Karin has created a couple of small bug/functionality issues in the alv repo, which I have started to take care of.

Thanks for your careful and constructive reviewing!
