### 2018-10-05 45de87f
Python has its own get_terminal_size in shutil and os from v3.3
