### 2018-10-10 #955 @karinlag
Checked off a few more, comfortable to move forward.
