### 2018-10-10 #955 @arvestad
Thanks @karinlag !
I'd just like to point out that the remaining non-checked item is about community guidelines and based on Karin's feedback I have added a file CONTRIBUTING.md (see https://github.com/arvestad/alv/blob/master/CONTRIBUTING.md) that I believe meets the requested needs.
