### 2018-10-11 86ded34
Started recording changes for v1.3 and changed the order to be 'most recent first'
