### 2018-10-16 ef4ba18
Added consideration of sorting sequences by leaf order.
