### 2018-11-05 b4a4b98
added changes for 1.2.1 release
