### 2018-11-05 0cdcb94
Added Zenodo badge with DOI
