### 2018-11-05 #955 @arvestad
Thanks a lot @yochannah . I have now put it on Zenodo and the DOI is 10.5281/zenodo.1477804.

Just to avoid confusion: I made a first erroneous Zenodo submission based on a development branch, so the DOI you see here is the second attempt. :-)
