### 2018-11-05 #955 @yochannah
@arvestad thanks! One small thing: in the Zenodo archive, there are two authors listed, but in the paper only one. You might need to edit the Zenodo metadata to remove any automatically added authors (I'm told this is a thing; I've never tried it myself so let me know how you get on!)
