### 2018-11-07 #955 @arfon
@arvestad - your paper is now accepted into JOSS :zap: :rocket: :boom:
