### 2018-11-07 #955 @arvestad
@arfon Hi, just a small note: the Markdown badge does not contain the right DOI. Should I file this as an issue somewhere? I edited it by hand, of course, so no big deal, but some users may not notice the error.
